feat : reservationService 예외처리 구문 추가

+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,6 +1,7 @@
 package roomescape.service;
 
 
+import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.dao.reservation.ReservationJdbcDAO;
@@ -20,14 +21,26 @@ public class ReservationService {
     }
 
     public Reservation createReservation(Reservation reservation) {
+        if (reservation.getDate().isBefore(LocalDate.now())) {
+            throw new IllegalArgumentException("Reservation date must be after today.");
+        }
         return reservationJdbcDAO.create(reservation);
     }
 
     public Reservation getReservationById(long id) {
+        Reservation reservation = reservationJdbcDAO.getById(id);
+        if (reservation == null) {
+            throw new IllegalArgumentException("No reservation found with the given ID.");
+        }
         return reservationJdbcDAO.getById(id);
     }
 
     public void deleteReservationById(long id) {
+        Reservation reservation = reservationJdbcDAO.getById(id);
+        if (reservation == null) {
+            throw new IllegalArgumentException("No reservation found with the given ID.");
+        }
         reservationJdbcDAO.delete(id);
     }
+
 }
